refactor: 위치 변경 되었을 때 지도 중심이동 로직을 control훅으로 이동 #187
